2023 clean-up

+++ strip-coverletter-docker.sh
@@ -33,12 +33,25 @@ chmod 777 ./vol
 function finish {
     retcode=$?
     if [[ $retcode -gt 0 ]]; then
-        # command has failed and a dump file was created
-        # move the dump file out of vol/ into the tmp dir
-        mv "vol/$bname_in.dump.tar" "$work_dir"
+        # failed
+    
         echo "failed: $retcode"
-        echo "wrote $work_dir/$bname_in.dump.tar"
-        echo "wrote $work_dir/$bname_in.log"
+        if [ -f "vol/$bname_in.dump.tar" ]; then
+            # command has failed and a dump file was created
+            # move the dump file out of vol/ into the work dir
+            mv "vol/$bname_in.dump.tar" "$work_dir"
+            echo "wrote $work_dir/$bname_in.dump.tar"
+        fi
+        echo "wrote $logfile"
+    else 
+        # success
+
+        # move final file to original destination
+        # mv -f "./vol/file.pdf" "/path/to/write/output/file.pdf"
+        mv -f "vol/$bname_out" "$out"
+
+        # remove log file
+        rm -f "$logfile"
     fi
 }
 trap finish EXIT
@@ -47,6 +60,7 @@ trap finish EXIT
 duration=1200
 timeout --preserve-status $duration \
     docker run \
+        --rm \
         --volume "$(pwd):/data" \
         --volume "$in:/data/$bname_in" \
         elifesciences/strip-coverletter "/data/$bname_in" "/data/vol/$bname_out" \
@@ -54,19 +68,11 @@ timeout --preserve-status $duration \
 
 # output files must be owned by the calling (host) user. got the idea here:
 # https://stackoverflow.com/questions/26500270/understanding-user-file-ownership-in-docker-how-to-avoid-changing-permissions-o/26514736#answer-54317162
-# lsh@2023-01-31: disabled, files are not being copied and to their destination and files appearing in /opt/elife-bot:
-# https://github.com/elifesciences/issues/issues/8088
-#docker run \
-#    --rm -it \
-#    --user root \
-#    --entrypoint "/bin/sh" \
-#    --env HOST_UID=$(id -u) \
-#    --volume "$(pwd)/vol:/data" \
-#    elifesciences/strip-coverletter -c 'chown -R ${HOST_UID}:${HOST_UID} /data/'
+docker run \
+    --rm -it \
+    --user root \
+    --entrypoint "/bin/sh" \
+    --env HOST_UID=$(id -u) \
+    --volume "$(pwd)/vol:/data" \
+    elifesciences/strip-coverletter -c 'chown -R ${HOST_UID}:${HOST_UID} /data/'
 
-# move final file to original destination
-# mv -f "vol/file.pdf" "/path/to/write/output/file.pdf"
-mv -f "vol/$bname_out" "$out"
-
-# remove log file
-rm -f "$logfile"
